add file class
